03-search-api/requirements.txt
langchain==0.3.2
langchain-community==0.3.1
tiktoken==0.8.0
sentence-transformers==3.1.1
einops==0.8.0
qdrant-client==1.11.3
fastapi==0.115.0
uvicorn==0.31.0
python-dotenv==1.0.1
